Better code for source file management and watching static files for changes
Note the +() syntax for referring to source file subdirectories in gulpfile.js - it
is necessary to preserve the directory structure in the output directory.

Introduction to path globs (as used in this context): https://github.com/isaacs/node-glob#glob-primer

The **/* pattern for catching "all files" in a subtree (versus just **) makes only a slight
difference, resulting in the exclusion of the base directory itself.

See: https://github.com/isaacs/minimatch#usage